move het hash to sign
